moved ObservationTimer
