+++ src/Controller/Controllers/GamePlayController.java
@@ -3,16 +3,13 @@ package Controller.Controllers;
 import Controller.Controller;
 import Controller.ControllerManager;
 import Controller.ControllerUtility.Command;
-import Model.Skills.Skill;
 import Model.State.GameState.ActiveGameState;
 import Utilities.GameLoaderSaver.GameSaver;
-import Utilities.ObservationTimer;
+import Utilities.Timer.ObservationTimer;
 import Utilities.Settings;
-import Utilities.Timer.TimedEvent;
 import View.ViewUtilities.Panels.GamePanel;
 
 import java.awt.event.KeyEvent;
-import java.util.ArrayList;
 
 /**
  * Created by Wimberley on 4/6/16.
